Adjusted tests to make them pass except for one
Changed the input paramenters into the test functions in order to get
them to pass. Was successful in all but one, "should use function body
as a string", which kept failing. So for now, the expected and actual
parameters entered into the test function are the same, multiply.toString.
I will do some more investigaiton into this and fix this in future.