Fixed typos in test expected values
Confirmation of the tests were not showing up in KoansRunner.html due
to syntax errors in the code. Specifically, the expected values.
There were misplaced closing brackets that were deleted in order to get
KoansRunner to run as expected and check the tests.
Expected values were also changed in order to get all the tests to pass
successfully.