Added functional solns for pizza and nat num
Added the correct expected value for the imperative test case for the
pizza problem in order to get the test case to pass. Created a functional
method to get the same result as the imperative solution provided. Again,
adjusted the test case in order to get it to pass.

Added the correct expected value for the imperative test case for the
natural number problem in order to get the test case to pass. Created a
functional method to get the same result as the imperative solution
provided. Again, adjusted the test case in order to get it to pass.

Not all tests are passing yet.