Added solns for ingredient count and extra credit Qs
Added the correct expected value for the ingredient count imperative
function. However, the functional problem needs work, as although it
passes the test, it is not a proper functional implementation and
needs further work.

Also completed the first two extra credit problems successfully, but
need to complete the final three.